Split terminal selection into server-client components
