+++ client/src/app/terminal/components/page-client.tsx
@@ -1,0 +1,46 @@
+"use client";
+import React, { useState } from "react";
+import { useRouter } from "next/navigation";
+import { Problem, TerminalDetails } from "colorblind/shared/lib/models/models";
+import { TerminalPicker } from "colorblind/shared/components/TerminalPicker";
+import styles from "../page.module.css";
+
+type Props = {
+  terminals: TerminalDetails[];
+};
+
+export default function StepOneClient({ terminals }: Props) {
+  const router = useRouter();
+  const [problem, setProblem] = useState<Problem>();
+  const [error, setError] = useState<string>();
+  const [pickedTerminalId, setPickedTerminalId] = useState<string>();
+
+  return (
+    <div className={styles.form}>
+      <form
+        onSubmit={(e) => {
+          e.preventDefault();
+          router.push(`/terminal/${pickedTerminalId}`);
+        }}
+      >
+        <h1>Select a terminal</h1>
+        <div>
+          <TerminalPicker
+            terminals={terminals}
+            onSubmit={(terminalId) => {
+              setPickedTerminalId(terminalId);
+            }}
+          />
+          <button
+            type="submit"
+            disabled={!pickedTerminalId || !!error || !!problem}
+          >
+            Submit
+          </button>
+        </div>
+        {problem ? JSON.stringify(problem) : null}
+        {error}
+      </form>
+    </div>
+  );
+}
